Modification page de connexion

+++ src/Entity/Produits.php
@@ -76,10 +76,11 @@ class Produits
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
+
 
     public function setPrice(float $price): static
     {
